+++ cli/src/command/mod.rs
@@ -15,7 +15,7 @@
 
 //! Define the commands supported by the CLI.
 
-use crate::{lookup_account, CommandError, CommandT, NetworkOptions, TxOptions};
+use crate::{lookup_key_pair, CommandError, CommandT, NetworkOptions, TxOptions};
 use itertools::Itertools;
 use radicle_registry_client::*;
 
@@ -23,6 +23,7 @@ use sp_core::crypto::Ss58Codec;
 use structopt::StructOpt;
 
 pub mod account;
+pub mod key_pair;
 pub mod org;
 pub mod other;
 pub mod project;
@@ -32,16 +33,16 @@ fn parse_account_id(data: &str) -> Result<AccountId, String> {
     Ss58Codec::from_ss58check(data)
         .map_err(|err| format!("{:?}", err))
         .or_else(|address_error| {
-            lookup_account(data)
-                .map(|account| account.public())
-                .map_err(|account_error| {
+            lookup_key_pair(data)
+                .map(|key_pair| key_pair.public())
+                .map_err(|key_pair_error| {
                     format!(
                         "
-    ! Could not parse an ss58 address nor find a local account with the given name.
+    ! Could not parse an ss58 address nor find a local key-pair with the given name.
     ⓘ Error parsing SS58 address: {}
-    ⓘ Error looking up account: {}
+    ⓘ Error looking up key-pair: {}
     ",
-                        address_error, account_error
+                        address_error, key_pair_error
                     )
                 })
         })